#174 - removed previous implementation (custom refresh control) - instead used setting a themable table view background - moved themable table view background implementation to UITableViewController Extension - removed duplicated table view background code - bug fix: safe area was wrong in this case, because view needs to be stretched to the edge
