+++ ownCloud/Server List/ServerListTableViewController.swift
@@ -213,23 +213,7 @@ class ServerListTableViewController: UITableViewController, Themeable {
 			self.navigationController?.navigationBar.shadowImage = UIImage()
 			self.tableView.tableHeaderView?.applyThemeCollection(Theme.shared.activeCollection)
 
-			// UITableView background view is nil for default. Set a UIView with clear color to can insert a subview above
-			let backgroundView = UIView.init(frame: self.tableView.frame)
-			backgroundView.backgroundColor = UIColor.clear
-			self.tableView.backgroundView = backgroundView
-
-			// This view is needed to stop flickering when scrolling (white line between UINavigationBar and UITableView header
-			let coloredView = ThemeableColoredView.init(frame: CGRect(x: 0, y: -self.view.frame.size.height, width: self.view.frame.size.width, height: self.view.frame.size.height + 1))
-			coloredView.translatesAutoresizingMaskIntoConstraints = false
-
-			self.tableView.insertSubview(coloredView, aboveSubview: self.tableView.backgroundView!)
-
-			NSLayoutConstraint.activate([
-				coloredView.topAnchor.constraint(equalTo: self.tableView.topAnchor, constant: -self.view.frame.size.height),
-				coloredView.leftAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.leftAnchor),
-				coloredView.rightAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.rightAnchor),
-				coloredView.heightAnchor.constraint(equalToConstant: self.view.frame.size.height + 1)
-			])
+			self.addThemableBackgroundView()
 		}
 	}
